기능 - 나눗셈 롱으로 수정

+++ Project5_18/app/src/main/java/com/example/project5_18/MainActivity.java
@@ -53,7 +53,7 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void div(View view) {
-        double value = Long.parseLong(first.getText().toString()) / Long.parseLong(second.getText().toString());
+        long value = Long.parseLong(first.getText().toString()) / Long.parseLong(second.getText().toString());
         result.setText("계산결과 : " + value);
     }
